core/gpu_sw_backend: drop redundant u8/u32 intermediate casts in rasteriser
DrawRectangle, DrawSpan, and DrawLine all contained two related patterns
that did the same thing as a single plain cast but with an extra layer:

  static_cast<u8>(static_cast<u32>(s32_y)) & 1u    -> y & 1
  static_cast<u8>(static_cast<u32>(u8_v) + u32_o)  -> static_cast<u8>(u8_v + u32_o)

For the interlace LSB check (3 sites: lines 231, 363, 686), `y` is s32,
the cast-to-u32 reinterprets bits (no-op for non-negative, two's-complement
preserving for negative), the cast-to-u8 truncates to low 8 bits, and the
`& 1u` keeps the LSB. Since the low bit of an integer is the same
regardless of which signed/unsigned width it's viewed through, the whole
ladder reduces to `y & 1`. The comparison against `active_line_lsb`
(BitField<u8,u8,1,1>::GetValue() returns u8) then promotes both sides to
int for the equality check - same as before.

For the texcoord wrap (2 sites: lines 236, 244), `origin_texcoord_y` is
u8 (from UnpackTexcoord -> std::tuple<u8,u8>) and `offset_y` is u32 (the
loop variable). The original `static_cast<u32>(origin_texcoord_y) +
offset_y` makes the u32 widening explicit; `origin_texcoord_y +
offset_y` does the same thing implicitly via the usual arithmetic
conversions (u8 integer-promotes to int, then to u32 to match the other
operand). Both end with the same `static_cast<u8>` narrow.

Equivalence verified with a small test harness covering boundary inputs
(0, 255, INT32_MIN, INT32_MAX, negative odd/even, wrap-around offsets);
all 27 cases produce identical results. -O2 g++ x86_64 emits byte-
identical assembly for both pairs:

  interlace check : andl $1, %esi ; movzbl %dil, %edi ; cmpl %edi, %esi
  texcoord wrap   : leal (%rdi,%rsi), %eax

-Wall and -Wconversion delta on the file: zero (13 -> 13 -Wconversion,
all pre-existing in unrelated lines).

Spotted as a follow-up after the helper fold; once `Truncate8` and
`ZeroExtend32` weren't hiding inside function calls, the redundancy
of the explicit double-cast chain became visible.

+++ src/core/gpu_sw_backend.cpp
@@ -228,12 +228,12 @@ void GPU_SW_Backend::DrawRectangle(const GPUBackendDrawRectangleCommand* cmd)
   {
     const s32 y = origin_y + static_cast<s32>(offset_y);
     if (y < static_cast<s32>(m_drawing_area.top) || y > static_cast<s32>(m_drawing_area.bottom) ||
-        (cmd->params.interlaced_rendering && cmd->params.active_line_lsb == (static_cast<u8>(static_cast<u32>(y)) & 1u)))
+        (cmd->params.interlaced_rendering && cmd->params.active_line_lsb == (y & 1)))
     {
       continue;
     }
 
-    const u8 texcoord_y = static_cast<u8>(static_cast<u32>(origin_texcoord_y) + offset_y);
+    const u8 texcoord_y = static_cast<u8>(origin_texcoord_y + offset_y);
 
     for (u32 offset_x = 0; offset_x < cmd->width; offset_x++)
     {
@@ -241,7 +241,7 @@ void GPU_SW_Backend::DrawRectangle(const GPUBackendDrawRectangleCommand* cmd)
       if (x < static_cast<s32>(m_drawing_area.left) || x > static_cast<s32>(m_drawing_area.right))
         continue;
 
-      const u8 texcoord_x = static_cast<u8>(static_cast<u32>(origin_texcoord_x) + offset_x);
+      const u8 texcoord_x = static_cast<u8>(origin_texcoord_x + offset_x);
 
       ShadePixel<texture_enable, raw_texture_enable, transparency_enable, false>(
         cmd, static_cast<u32>(x), static_cast<u32>(y), r, g, b, texcoord_x, texcoord_y);
@@ -360,7 +360,7 @@ template<bool shading_enable, bool texture_enable, bool raw_texture_enable, bool
 void GPU_SW_Backend::DrawSpan(const GPUBackendDrawPolygonCommand* cmd, s32 y, s32 x_start, s32 x_bound, i_group ig,
                               const i_deltas& idl)
 {
-  if (cmd->params.interlaced_rendering && cmd->params.active_line_lsb == (static_cast<u8>(static_cast<u32>(y)) & 1u))
+  if (cmd->params.interlaced_rendering && cmd->params.active_line_lsb == (y & 1))
     return;
 
   s32 x_ig_adjust = x_start;
@@ -683,7 +683,7 @@ void GPU_SW_Backend::DrawLine(const GPUBackendDrawLineCommand* cmd, const GPUBac
     const s32 x = (cur_point.x >> Line_XY_FractBits) & 2047;
     const s32 y = (cur_point.y >> Line_XY_FractBits) & 2047;
 
-    if ((!cmd->params.interlaced_rendering || cmd->params.active_line_lsb != (static_cast<u8>(static_cast<u32>(y)) & 1u)) &&
+    if ((!cmd->params.interlaced_rendering || cmd->params.active_line_lsb != (y & 1)) &&
         x >= static_cast<s32>(m_drawing_area.left) && x <= static_cast<s32>(m_drawing_area.right) &&
         y >= static_cast<s32>(m_drawing_area.top) && y <= static_cast<s32>(m_drawing_area.bottom))
     {
